fix(Serial_Manager): Singleton a. add more comment
	Serial_Manager is now Singleton
	Add more comments to Serial Source File
	change Sequenznumber from int to uint8_t

+++ project/ESEP/src/lib/serial/paket_protocol.h
@@ -30,7 +30,7 @@ enum PAKET_TYP {
 typedef struct {
 	size_t paket_size;
 	PAKET_TYP paket_typ;
-	int sequenznummer;
+	uint8_t sequenznummer;
 } Header;
 
 /**
